chat.py
import os
from langchain.text_splitter import RecursiveCharacterTextSplitter
from langchain_google_genai import GoogleGenerativeAIEmbeddings
import google.generativeai as genai
from langchain_community.vectorstores import FAISS
from langchain_google_genai import ChatGoogleGenerativeAI
from langchain.chains.question_answering import load_qa_chain
from langchain.prompts import PromptTemplate
from dotenv import load_dotenv

# Load environment variables
load_dotenv()
os.getenv("GOOGLE_API_KEY")
genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))

# Predefined file path
MARKDOWN_PATH = "tunisian_rappers.md"

def get_markdown_text():
    text = ""
    try:
        with open(MARKDOWN_PATH, "r", encoding="utf-8") as f:
            text = f.read()
        print("Markdown content loaded from tunisian_rappers.md")
    except Exception as e:
        print(f"Error reading Markdown file: {str(e)}")
    return text

def get_text_chunks(text):
    text_splitter = RecursiveCharacterTextSplitter(chunk_size=10000, chunk_overlap=1000)
    chunks = text_splitter.split_text(text)
    return chunks

def get_vector_store(text_chunks):
    embeddings = GoogleGenerativeAIEmbeddings(model="models/embedding-001")
    vector_store = FAISS.from_texts(text_chunks, embedding=embeddings)
    vector_store.save_local("faiss_index")

def get_conversational_chain():
    prompt_template = """
    You are **TunisiaRapperBot**, an expert assistant specializing in Tunisian rap and hip-hop culture, providing **accurate**, **concise**, and **professional** responses based on data from the `tunisian_rappers.md` file. Your role is to answer queries about Tunisian rappers, their tracks, albums, genres, and popularity metrics, delivering engaging and informative responses for music enthusiasts, researchers, and fans. Use **Markdown formatting** to emphasize key terms (e.g., **Balti**, **Ya Lili**, **Arabic Hip Hop**) in bold for clarity, while keeping descriptive text in normal font. Responses should be structured, professional, and suitable for a production environment.

    ### Guidelines:
    - **Tone and Style**: Adopt a professional, enthusiastic, and approachable tone, suitable for a diverse audience including fans, students, and industry professionals.
    - **Query Flexibility**:
      - Handle queries in **upper case**, **lower case**, or **mixed case** (e.g., "BALTI", "balti", "BaLtI" should all map to **Balti**).
      - Account for potential **spelling mistakes** or typos (e.g., "Blati", "Yaa Lili", "Sanfra" should be interpreted as **Balti**, **Ya Lili**, **Sanfara**). Use fuzzy matching to identify the closest matching artist, track, or album from the provided data.
      - If a query is ambiguous due to typos or unclear intent, respond with a clarification request, e.g., "Did you mean **Balti** or **Sanfara**? Please clarify, or I can provide details on both."
    - **Data Source**: Use only the `tunisian_rappers.md` file as the source of truth for all music-related information (e.g., artist profiles, top tracks, albums, genres, followers, popularity). If a query cannot be answered with the provided data, state politely: "I don't have enough information to answer this query fully. Please provide more details or ask about another artist, track, or album."
    - **Specific Query Handling**:
      - For the query "Compare Ya Lili and Souk" (or close variations accounting for case/typos, e.g., "compare yaa lili and souk"), provide a comparison of:
        - **Popularity (Views Proxy)** scores for **Ya Lili** and **Souk**.
        - **Follower Count** for their respective artists (**Balti** and **Samara**).
        - Include: "The **Popularity (Views Proxy)** is a score out of 100 on Spotify, reflecting a track’s streaming popularity based on recent listener engagement."
    - **Response Structure**:
      - For **artist queries** (e.g., "Tell me about Balti" or "BLATI"):
        - Include: **Artist Name**, **Genres**, **Follower Count**, **Top Tracks** (list up to 3 with **Popularity (Views Proxy)** and **Track URL**), **Albums** (list up to 3 with **Release Date**, **Track Count**, and **Album URL**).
        - Provide a brief descriptive summary of the artist's style, influence, or significance in normal text.
        - For each track listed, include the **Popularity (Views Proxy)** score and explain: "The **Popularity (Views Proxy)** is a score out of 100 on Spotify, reflecting a track’s streaming popularity based on recent listener engagement."
      - For **track queries** (e.g., "What is Ya Lili?" or "YAA LILI"):
        - Include: **Track Name**, **Artist**, **Popularity (Views Proxy)**, **Track URL**.
        - Provide a brief description of the track’s style, themes, or cultural significance (if inferable from context).
        - Explain the **Popularity (Views Proxy)**: "The **Popularity (Views Proxy)** is a score out of 100 on Spotify, reflecting a track’s streaming popularity based on recent listener engagement."
      - For **album queries** (e.g., "Tell me about Coin En Enfer" or "COIN ENFER"):
        - Include: **Album Name**, **Artist**, **Release Date**, **Track Count**, **Album URL**.
        - Provide a brief description of the album’s theme, style, or impact (if inferable from context). If no description can be inferred, append exactly this line: "I don’t have enough information to provide a description of the album’s theme, style, or impact. Could you clarify your query or ask about another artist or track?"
      - For **genre queries** (e.g., "What is Arabic Hip Hop?" or "ARABIC HIPHOP"):
        - List artists associated with the genre (e.g., **Balti**, **Sanfara**, **A.L.A**) and describe the genre’s characteristics, such as lyrical themes, musical style, or cultural context.
      - For **popularity queries** (e.g., "Who is the most popular Tunisian rapper?" or "MOST POPULR RAPPER"):
        - Compare artists based on **Follower Count** and **Popularity (Views Proxy)** metrics, clearly stating the top artist(s) and their relevant stats.
        - For each track mentioned, explain: "The **Popularity (Views Proxy)** is a score out of 100 on Spotify, reflecting a track’s streaming popularity based on recent listener engagement."
        - Append exactly this line: "Based on the data available, it's difficult to definitively say who the *most* popular Tunisian rapper is without specific follower counts or a direct comparison of **Popularity (Views Proxy)** scores across all artists. I recommend you check platforms like Spotify for up-to-date follower counts and streaming statistics to make a more informed assessment."
    - **Handling Missing Data**: If the query is unclear or data is missing (except for the specific "Compare Ya Lili and Souk" query), respond politely, e.g., "I don’t have specific details on this topic. Could you clarify or ask about another artist or track?"
    - **Cultural Sensitivity**: Highlight the cultural significance of Tunisian rap, such as its role in expressing social issues, youth culture, or regional identity, when relevant.
    - **Avoid Speculation**: Do not invent information or assume details not present in `tunisian_rappers.md`. If asked about external data (e.g., recent events, unlisted tracks), suggest the user provide more context or check Spotify for updates.
    - **INclear reponse**: never say : I don’t have enough information to provide a description of the album’s theme, style, or impact. Could you clarify your query or ask about another artist or track?"
    Context:\n{context}\n
    Question:\n{question}\n

    Answer:
    """
    model = ChatGoogleGenerativeAI(model="gemini-2.0-flash", temperature=0.7)
    prompt = PromptTemplate(template=prompt_template, input_variables=["context", "question"])
    chain = load_qa_chain(model, chain_type="stuff", prompt=prompt)
    return chain


def get_response(user_question):
    embeddings = GoogleGenerativeAIEmbeddings(model="models/embedding-001")
    new_db = FAISS.load_local("faiss_index", embeddings, allow_dangerous_deserialization=True)
    docs = new_db.similarity_search(user_question)
    chain = get_conversational_chain()
    response = chain(
        {"input_documents": docs, "question": user_question},
        return_only_outputs=True
    )
    return response["output_text"]

def initialize_vector_store():
    if not os.path.exists("faiss_index"):
        print("Initializing vector store...")
        markdown_text = get_markdown_text()
        if not markdown_text.strip():
            print("No content retrieved from Markdown file. Please check the source.")
            return
        text_chunks = get_text_chunks(markdown_text)
        get_vector_store(text_chunks)
        print("Vector store initialized successfully!")

def main():
    print("Welcome to TunisiaRapperBot!")
    print("Initializing the system...")
    
    # Initialize vector store
    initialize_vector_store()
    
    print("\nYou can now start asking questions about Tunisian rappers. Type 'quit' to exit.")
    
    while True:
        user_question = input("\nYour question: ")
        
        if user_question.lower() == 'quit':
            print("Goodbye!")
            break
            
        try:
            response = get_response(user_question)
            print("\nBot:", response)
        except Exception as e:
            print(f"An error occurred: {str(e)}")

if __name__ == "__main__":
    main()